fix: parseNpass adapted to sync
